backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
bcrypt
python-dotenv
requests
Pillow
django-cors-headers